refact divid mvp value
